fixed 지도 페이지 분리 => 게시글 작성시 지도 위치가 반영은 되나 리렌더링이 일어나지 않아 주소, 좌표가 리덕스로 가지 않던 문제 수정
