--- src/pages/Map.tsx
@@ -1,8 +1,0 @@
-import React, { FC } from 'react';
-import MapContainer from '../components/MapContainer';
-
-const Map: FC = () => {
-  return <MapContainer />;
-};
-
-export default Map;
